+++ src/GitHub_Updater/Bitbucket_API.php
@@ -36,7 +36,7 @@ class Bitbucket_API extends API implements API_Interface {
 		$this->type     = $type;
 		$this->response = $this->get_repo_cache();
 
-		$this->load_authentication_hooks();
+		Basic_Auth_Loader::instance( parent::$options )->load_authentication_hooks();
 
 		if ( ! isset( self::$options['bitbucket_username'] ) ) {
 			self::$options['bitbucket_username'] = null;
@@ -45,22 +45,6 @@ class Bitbucket_API extends API implements API_Interface {
 			self::$options['bitbucket_password'] = null;
 		}
 		add_site_option( 'github_updater', self::$options );
-	}
-
-	/**
-	 * Load hooks for Bitbucket authentication headers.
-	 */
-	public function load_authentication_hooks() {
-		add_filter( 'http_request_args', array( &$this, 'maybe_basic_authenticate_http' ), 5, 2 );
-		add_filter( 'http_request_args', array( &$this, 'http_release_asset_auth' ), 15, 2 );
-	}
-
-	/**
-	 * Remove hooks for Bitbucket authentication headers.
-	 */
-	public function remove_authentication_hooks() {
-		remove_filter( 'http_request_args', array( &$this, 'maybe_basic_authenticate_http' ) );
-		remove_filter( 'http_request_args', array( &$this, 'http_release_asset_auth' ) );
 	}
 
 	/**